requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
google-generativeai
langgraph
pypdf
faiss-cpu
pydantic
python-dotenv